fix: delay fcu to 500ms before the next slot
