+++ packages/beacon-node/src/chain/interface.ts
@@ -211,6 +211,9 @@ export interface IBeaconChain {
 
   recomputeForkChoiceHead(caller: ForkchoiceCaller): ProtoBlock;
 
+  /** Issue fcu call to EL */
+  notifyForkchoiceUpdate(): Promise<void>;
+
   /** When proposerBoostReorg is enabled, this is called at slot n-1 to predict the head block to build on if we are proposing at slot n */
   predictProposerHead(slot: Slot): ProtoBlock;
